Refactor BookController to use LocalDbContext
The BookController has been refactored to use LocalDbContext instead of a json file for persisting book details. The Books model has also been modified to reflect this change, and UserController has been updated with Patch and Delete methods. This result in a more standardized way of manipulating data across controllers.